seeders admins roles ...

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -22,7 +22,8 @@ return new class extends Migration
             $table->string('first_name')->nullable()->comment('Имя');
             $table->string('last_name')->nullable()->comment('Фамилия');
             $table->string('middle_name')->nullable()->comment('Отчество');
-            $table->unsignedBigInteger('phone_number')->nullable()->comment('Телефон номер в полном формате без знаков');
+            $table->decimal('phone_number')->nullable()
+                ->comment('Телефон номер в полном формате без знаков');
             $table->dateTime('birthday')->nullable()->comment('Дата рождение');
 
             $table->boolean('active')->default(true)->comment('Активность');
